Accordion in Flutter
1. Add dependencies:
  accordion: ^2.5.1 to your flutter project in pubspec.yaml

2. Import these files to your lib and replace your main.dart file

3. $flutter run

Follow @alfawhocodes for more Flutter App development
tips
.
.
.
.
.

FAQ’s

What is Flutter?'
Flutter SDK (Software Development Kit) is an innovative technology for creating applications on mobile, desktop, and web platforms. It’s a free, open-source UI framework from Google which has hit the market in 2017.

 Is a Flutter app performance as good as native app?
Yes, you can say so.If you don’t need to use any system-specific features, or latest technology solutions in your app, you wouldn’t see any difference in performance whether it\’s created with Flutter or natively. The difference is that with Flutter, you can get it done faster.

Is Flutter a good choice for startups?
Absolutely! Startups owners knowing the advantages of Flutter choose for their mobile app development more and more often. It’s a great solution for those who want to release the app as fast as possible to test the idea and the product, if it has reached the target audience and if it can be used to make a profit. Flutter apps can be developed further anytime later.

Is Flutter good for MVP development?
Yes. As mentioned above, companies often use Flutter for their mobile app development to test and validate the idea, which is what MVPs are all about. Flutter apps can be released faster than native apps, that’s why it’s a good choice for MVPs as well.

What are some popular Flutter apps?
Alibaba
- Google Pay
- Google Ads
- Ad Stats for AdMob
- Reflectly
- eBay
- iRobot

 Want to learn Flutter ?
 Do follow @AlfaWhocCodes
 DM 'FLUTTER' for mentorship